requirements.txt
click==8.0.1
llvmlite==0.36.0
numba==0.53.1
numpy==1.20.3
Pillow==8.2.0
scipy==1.7.0
